refactor(admin): extract Settings > SEO + sitemap + structured data into standalone component
Move the cmsAdvanced-gated SEO panel (meta title/description per page+language,
sitemap preview, structured-data validation) out of the monolithic
AdminComponent into a new standalone app-admin-seo (OnPush). The panel owns its
own SEO form and sitemap/structured-data result state; seoPage stays shared with
the settings content-revisions selector via two-way binding, and the CMS
content-version helpers are threaded in through the existing bound-callback
inputs. Behaviour-preserving: template moved verbatim and the panel's unit tests
move to admin-seo.component.spec.ts, with the rendered-DOM net flow now exercising
the extracted component.
